requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250706
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blake3==1.0.5
blobfile==3.0.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.7
compressed-tensors==0.10.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cupy-cuda12x==13.5.1
datasets==4.0.0
decorator==5.2.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email-validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastrlock==0.8.3
filelock==3.18.0
flash-attn==2.7.4.post1
flashinfer-python==0.2.5
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
gitdb==4.0.12
gitpython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
hydra-core==1.3.2
identify==2.6.12
idna==3.10
importlib-metadata==8.7.0
interegular==0.3.3
ipython==8.37.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lark==1.2.2
latex2sympy2-extended==1.10.2
liger-kernel==0.6.0
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==6.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
math-verify==0.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral-common==1.8.1
modelscope==1.28.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.90.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-fixer==0.2.5
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
opentelemetry-api==1.35.0
opentelemetry-exporter-prometheus==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
optree==0.16.0
orjson==3.11.0
outlines==0.1.11
outlines-core==0.1.26
packaging==25.0
pandas==2.3.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
peft==0.16.0
pexpect==4.9.0
pillow==11.3.0
# pip @ file:///croot/pip_1746204010231/work
platformdirs==4.3.8
pre-commit==4.2.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.1
pybind11==3.0.0
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pygments==2.19.2
pylatexenc==2.10
pynvml==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyvers==0.1.0
pyyaml==6.0.2
pyzmq==27.0.0
ray==2.48.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.33.0
setproctitle==1.3.6
setuptools==78.1.1
sgl-kernel==0.2.6
sglang==0.4.6.post5
shellingham==1.5.4
six==1.17.0
smart-open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
stack-data==0.6.3
starlette==0.47.1
sympy==1.14.0
tensordict==0.9.1
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.0
torch-memory-saver==0.0.8
torchao==0.9.0
torchaudio==2.7.0
torchdata==0.11.0
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triton==3.3.0
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.0
uvicorn==0.35.0
uvloop==0.21.0
# -e file:///home/nfs06/chenyq/verl
virtualenv==20.31.2
vllm==0.9.2
wandb==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xformers==0.0.30
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zmq==0.0.0
